Choose first OSGi jar found
In the case that the Eclipse installation has been updated
multiple OSGi jars will be found, causing the java -classpath
command line argument to be incorrect as it will have a space
in it.

This finds the OSGi jar the same way as the other jars
further down the file.